+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @storybook-gabriel-ui/tokens
 
+## 2.2.0
+
+### Minor Changes
+
+- Addyng new components on Design system: Toast component and Tooltip Component
+
 ## 2.1.0
 
 ### Minor Changes